Renamed and put into correct namespaces
