+++ Turnbased-Game/Models/Client/IHost.cs
@@ -1,6 +1,6 @@
-namespace DefaultNamespace;
+namespace Turnbased_Game.Models.Client;
 
-public interface Host
+public interface IHost
 {
     
     public void ChangeGameSettings(string settings); // JSON
