+++ Turnbased-Game/Models/Client/IPlayer.cs
@@ -1,0 +1,12 @@
+namespace Turnbased_Game.Models.Client;
+
+public interface IPlayer
+{
+    public void SubmitTurn(string turn);
+    
+    public void SendMessage(int playerId, string message);
+    public void RequestRoleChange(IRole role);
+    public void RequestProfileUpdate(IPlayerProfile profile);
+    public void IsReady();
+    public void IsNotReady();
+}
