--- Turnbased-Game/Models/Game.cs
@@ -1,6 +1,0 @@
-namespace DefaultNamespace;
-
-public interface Game
-{
-    
-}
