+++ Turnbased-Game/Models/IGame.cs
@@ -1,0 +1,6 @@
+namespace Turnbased_Game.Models;
+
+public interface IGame
+{
+    
+}
