--- Turnbased-Game/Models/PlayerProfile.cs
@@ -1,6 +1,0 @@
-namespace DefaultNamespace;
-
-public interface PlayerProfile
-{
-    
-}
