--- Turnbased-Game/Models/Role.cs
@@ -1,6 +1,0 @@
-namespace DefaultNamespace;
-
-public interface Role
-{
-    
-}
